modificando:  modified:   rest_api/serializers.py, modified:   rest_api/urls.py, modified:   rest_api/views.py
