+++ rest_api/views.py
@@ -1,21 +1,38 @@
-from django.shortcuts import render
-from rest_framework.decorators import api_view
-from rest_framework.response import Response
-from rest_framework.viewsets import ModelViewSet
-from reserva.models import Reserva
-from rest_api.serializers import AgendamentoModelSerializer
-from rest_framework.authentication import TokenAuthentication
-from rest_framework.permissions import IsAuthenticated
-from rest_framework.permissions import IsAuthenticatedOrReadOnly
+#from django.shortcuts import render
+#from rest_framework.decorators import api_view
+#from rest_framework.response import Response
 
 
-class AgendamentoModelViewSet(ModelViewSet):
-    queryset = Reserva.objects.all()
-    serializer_class = AgendamentoModelSerializer
-    authentication_classes = [TokenAuthentication]
-    permission_classes = [IsAuthenticatedOrReadOnly]
+ 
+
+from rest_framework.serializers import ModelSerializer, HyperlinkedRelatedField
+
+from .serializers import AgendamentoModelSerializer
+from reserva.models import  Petshop
+
+from rest_framework.serializers import ModelSerializer, HyperlinkedRelatedField
+
+from .models import Petshop
 
 
+
+
+class PetshopModelSerializer(ModelSerializer):
+    reservas = HyperlinkedRelatedField(
+        many=True,
+        read_only=True,
+        view_name='api: reserva-detail'
+
+    )
+
+    class Meta: 
+        model = Petshop
+        fields = '__all__'
+
+
+
+
+'''
 @api_view(['GET', 'POST', 'PUT'])
 def hello_world(request):
     if request.method == 'POST':
@@ -26,3 +43,4 @@ def hello_world(request):
     
     else:
         return Response ({'Meu Saldo': 'R$5.110,200 To Rico $$$$$$!Ha!Ha!'})
+'''
